backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# MinIO
minio==7.2.3
httpx==0.26.0

# Celery
celery[redis]==5.3.6
vine==5.1.0

# LLM
openai==1.12.0
anthropic==0.18.0
aiohttp==3.9.3

# Config
pyyaml==6.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
jinja2==3.1.3
weasyprint==60.2
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.1
aiofiles==23.2.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
